requirements-dev.txt
black==20.8b1
flake8==3.8.4
mypy==0.800
pytest==6.2.2
pytest-django==4.1.0
